make sure to add some db limit when querying all observations
This has very little impact on performance, but it could once we have a
lot of obserations in the database.
In the previous implementation, all observations would be rendered into a
list. That list would be joined with all nests. The joined list
would be sorted and the limit would be applied.
In the new implementation, the limit will already be applied on both
queries (individuals and nests) which can make a difference.