Make sure also latest observations are shown on map
The time slider has a max time and uses intervals of one week. It seems that you can only
drag the slider to a point between the sliders boundaries (set by min and max) that is
exactly n times the interval.
To overcome this issue, I added 1 interval duration to the end property of the time
slider.
This was added as a comment to #276.